🚚 Move components from shared module to right place
