+++ frontend/src/app/pets/components/pets-card/pets-card.component.html
