+++ frontend/src/app/pets/components/pets-card/pets-card.component.ts
@@ -1,6 +1,6 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { Pet } from '../../../../pets/models/pet';
-import { PetsService } from '../../../../pets/services/pets.service';
+import { PetsService } from '../../services/pets.service';
+import { Pet } from '../../models/pet';
 
 @Component({
   selector: 'pets-card',
