+++ frontend/src/app/pets/components/pets-list/pets-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core';
-import { Pet } from '../../../../pets/models/pet';
+import { Pet } from '../../models/pet';
 
 @Component({
   selector: 'pets-list',
