+++ frontend/src/app/pets/pets.module.ts
@@ -3,8 +3,8 @@ import { SharedModule } from '../shared/shared.module';
 import { PetsAddComponent } from './components/pets-add/pets-add.component';
 import { PetsPageComponent } from './pages/pets-page/pets-page.component';
 import { PetsService } from './services/pets.service';
-import { PetsCardComponent } from '../shared/components/pets/pets-card/pets-card.component';
-import { PetsListComponent } from '../shared/components/pets/pets-list/pets-list.component';
+import { PetsListComponent } from './components/pets-list/pets-list.component';
+import { PetsCardComponent } from './components/pets-card/pets-card.component';
 
 @NgModule({
   declarations: [
