+++ frontend/src/app/reservations/components/reservation-hours-card/reservation-hours-card.component.ts
